ok, ok, put back a predict.py
